# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vault API
  description: >
    Osobní hudební systém — REST API. Pokrývá globální katalog/vyhledávání
    (MusicBrainz/Deezer proxy), asynchronní on-demand provisioning chybějících
    skladeb a doporučení/playlisty (ListenBrainz). Real-time synchronizace
    přehrávání je popsaná odděleně v `asyncapi.yaml`.
  version: "0.1.0"
servers:
  - url: https://vault.local/api/v1
    description: Přes reverse proxy / VPN (Tailscale), nikdy veřejně
security:
  - bearerAuth: []
...
paths:
...
  /tracks/{recordingId}/provision:
    post:
      tags: [provisioning]
      summary: >
        Idempotentní trigger zpřístupnění skladby. Pokud je MediaAsset již
        AVAILABLE, rovnou vrací stream info. Jinak založí (nebo najde
        existující) ProvisioningJob a vrátí jeho ID — klient dál poslouchá
        WS event `track.available`, tento endpoint se nevolá opakovaně v
        polling smyčce.
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          description: Skladba je již dostupná — přímý stream.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProvisionResult" }
        "202":
          description: Skladba chybí, byl založen/nalezen provisioning job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProvisionResult" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Skladbu nelze provisionovat (např. trvale FAILED bez retry politiky).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /library/local-tracks:
    get:
      tags: [library]
      summary: >
        Nahrávky naskenované z lokální knihovny (POST /library/scan) --
        vždy AVAILABLE bez provisioningu, rovnou přehratelné.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100, minimum: 1, maximum: 500 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Recording" }
...
  /library/by-genre/{genre}:
    get:
      tags: [library]
      summary: Lokální skladby daného MusicBrainz žánru (viz /library/genres).
      parameters:
        - name: genre
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Recording" }
...
  /library/import/spotify:
    post:
      tags: [library]
      summary: >
        Naimportuje Spotify export -- autodetekce podle obsahu, ne přípony:
        ZIP s CSV playlisty (Exportify apod., jeden soubor na playlist,
        `Liked_Songs.csv` jde do Liked Songs, ostatní se stanou vlastními
        playlisty) nebo oficiální `YourLibrary.json` (vždy do Liked Songs).
        Páruje skladby na katalog jménem, ne přes MusicBrainz.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalInFile: { type: integer }
                  matched: { type: integer }
                  alreadyPresent: { type: integer }
                  skipped: { type: integer }
                  playlistsImported: { type: integer }
        "400": { description: Nerozpoznaný formát (ani ZIP, ani JSON) }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Device-scoped JWT vydaný při párování zařízení.
  parameters:
...
    RecordingId:
      name: recordingId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Zdroj nenalezen
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    Availability:
      type: string
      enum: [available, provisionable, unavailable]
      description: >
        available = MediaAsset.status==AVAILABLE; provisionable = existuje
        v katalogu a lze založit provisioning job; unavailable = žádný
        známý zdroj (čistě metadatová položka bez cesty k obstarání).
...
    Recording:
      type: object
      properties:
        id: { type: string, format: uuid }
        mbid: { type: string, nullable: true }
        releaseId: { type: string, format: uuid, nullable: true }
        artistId: { type: string, format: uuid }
        title: { type: string }
        durationMs: { type: integer }
        isrc: { type: string, nullable: true }
        trackNumber: { type: integer, nullable: true }
        availability: { $ref: "#/components/schemas/Availability" }
        previewUrl:
          type: string
          format: uri
          nullable: true
          description: 30s náhled z Deezer (doplňkové metadata, best-effort — nemusí být vyplněno).
        listenCount:
          type: integer
          nullable: true
          description: >
            Počet poslechů z veřejného ListenBrainz API (doplňkové metadata,
            jen u /recommendations/trending a /recommendations/community —
            jinde vždy null).
...
    ProvisioningJob:
      type: object
      properties:
        id: { type: string, format: uuid }
        recordingId: { type: string, format: uuid }
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        attempts: { type: integer }
        createdAt: { type: string, format: date-time }
        errorMessage: { type: string, nullable: true }
    ProvisionResult:
      type: object
      required: [recordingId]
      properties:
        recordingId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/MediaAsset/properties/status" }
        streamUrl:
          type: string
          format: uri
          nullable: true
          description: Vyplněno pouze pokud status==AVAILABLE.
        job:
          allOf: [{ $ref: "#/components/schemas/ProvisioningJob" }]
          nullable: true
          description: Vyplněno pouze pokud se čeká na obstarání (HTTP 202).
...